Added NetworkTensor class.
Like a TreeTensor but without the requirement of being a tree. This means that contractions cannot be guaranteed to be efficient, so this class should only be used either as a base class (e.g. TreeTensor could be made a subclass) or in cases where it is known that the network can be efficiently contracted by naive methods.